sitekit/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
#certifi==2025.11.12
#cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.8
django-debug-toolbar==4.2.0
django-extensions==4.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
executing==2.2.1
hiredis==3.3.0
idna==3.11
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.3.1
parso==0.8.5
pillow==12.0.0
prompt_toolkit==3.0.52
#psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python3-openid==3.2.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13

gevent==24.11.1
gunicorn==23.0.0